ran
holy shit so the problem WAS FUCKING OFFSET WASN'T SET TO DISCRETE FFFFFFFFFUCK but now I get it.

I tried to align the sprites as they were in MUGEN rather than align them to a back or feet, it seems to look like a proper run but still looks a little jank.